Python/python-trojan exhibit/Python/Trojan.Python.AngstStealer/requirements.txt
pycryptodome==3.9.8
requests==2.24.0
pyinstaller==3.6
pyarmor==6.3.3
pywin32==227
